pyproject.toml
[tool.poetry]
name = "sleep-edf"
version = "0.1.0"
description = ""
authors = ["Gihyuk Kwon <imrlgur81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyedflib = "^0.1.37"
matplotlib = "^3.9.0"
pandas = "^2.2.2"
numpy = "1.26.3"
scikit-learn = "^1.5.1"
xgboost = "^2.1.0"
catboost = "^1.2.5"
pyspark = "^3.5.1"
dask = {extras = ["dataframe"], version = "^2024.7.0"}
hydra-core = "^1.3.2"
mlflow = "^2.14.2"
imbalanced-learn = "^0.12.3"
plotly = "^5.22.0"
nbformat = ">=4.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
